requirements.txt
Django>=2.0,<3.0
mysqlclient
mysql-connector-python
djangorestframework
django-cors-headers
django-url-filter